#1001: Refactored out cslc_utils to break out cslc dependency and blackout dates functionalities into their own modules. It was getting a bit unwieldy
