+++ data_subscriber/query.py
@@ -14,12 +14,12 @@ from data_subscriber.cmr import (async_query_cmr,
                                  ProductType, DateTimeRange,
                                  COLLECTION_TO_PRODUCT_TYPE_MAP,
                                  COLLECTION_TO_PROVIDER_TYPE_MAP)
-from data_subscriber.cslc_utils import split_download_batch_id
 from data_subscriber.geojson_utils import (localize_include_exclude,
                                            filter_granules_by_regions,
                                            download_from_s3)
 from data_subscriber.rtc.rtc_download_job_submitter import submit_rtc_download_job_submissions_tasks
-from data_subscriber.cslc_utils import split_download_batch_id, save_blocked_download_job, CSLCDependency
+from data_subscriber.cslc_utils import split_download_batch_id, save_blocked_download_job
+from data_subscriber.cslc.cslc_dependency import CSLCDependency
 from data_subscriber.url import form_batch_id, _slc_url_to_chunk_id
 from hysds_commons.job_utils import submit_mozart_job
 from util.conf_util import SettingsConf
